syntax correction

+++ lib/addSub.c
@@ -1,11 +1,11 @@
-# include <header.h>
+# include <stdio.h>
 
-int add(int num1, int num2)
+double add(double num1, double num2)
 {
 	return (num1+num2);
 }
 
-int sub(int num1, int num2)
+double sub(double num1, double num2)
 {
 	return (num1-num2);
 }
